Avoid nested include of <vespa/fastos/file.h>
